Alerta eliminar yy confirmacion

+++ controllers/AdminController.php
@@ -7,7 +7,9 @@ use Models\Derechohabiente;
 
 class AdminController{
 
-    public static function update(){}
+    public static function update(){
+      
+    }
     public static function delete(){
  
         $derechoHabiente = new Derechohabiente;
@@ -20,28 +22,31 @@ class AdminController{
         $sentencia = "curpD='" . $curp . "'";
 
         //Obtenemos la boleta para poder borrar el pdf y la foto asociados
-      //  $childBoleta = $child->some($sentencia)[0];
+       $childBoleta = $child->some($sentencia)[0];
         
         //Borrado del hijo y conyuguE
         $child->delete($sentencia);
 
         //verificar que tenga conyugue
-      // $exiteConyu = $conyugue->some($sentencia);
+       $exiteConyu = $conyugue->some($sentencia);
 
        //Si lo tiene lo borramos
-       /*if($exiteConyu){
+       if($exiteConyu){
         $conyugue->delete($sentencia);
-       }*/
+
+        //Borra imagen del conyugue
+        unlink(__DIR__ . '/../fotos/' . $curp . 'Con.jpg');
+       }
         //Borrado del padre
         $derechoHabiente->delete($sentenciaDerecho);
 
-      /*  //Borrado de pdf
+        //Borrado de pdf
          unlink(__DIR__ . '/../generadosPdf/' . $childBoleta->boleta . '.pdf');
         //Borrado de fotos
         //child
-        unlink(__DIR__ . '/../fotos/' . $childBoleta->boleta . 'jpg');
+        unlink(__DIR__ . '/../fotos/' . $childBoleta->boleta . '.jpg');
         //Padre
-        unlink(__DIR__ . '/../fotos/' . $curp . 'jpg');*/
+        unlink(__DIR__ . '/../fotos/' . $curp . '.jpg');
         //Le mandamos una respuesta el front
         echo json_encode(['respuestaB' => 1]);
     }
